code/requirements.txt
python-telegram-bot
pythonping
pyyaml